# blackend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TAST BOARD",
    "description": "Description",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "LOCAL SERVER"
    },
    {
      "url": "http://TEST",
      "description": "TEST SERVER"
    },
    {
      "url": "http://PROD",
      "description": "PROD SERVER"
    }
  ],
  "paths": {
    "/api/board/task": {
      "get": {
        "tags": [
          "Todo"
        ],
        "summary": "GET ALL TODO LIST",
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Todo"
        ],
        "summary": "ADD NEW TODO IN LIST",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Add in todo list",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Hello"
                },
                "description": {
                  "type": "string",
                  "example": "this is working today"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Todo"
        ],
        "summary": "EDIT TODO IN LIST",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Add in todo list",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "uuid"
                },
                "title": {
                  "type": "string",
                  "example": "Hello"
                },
                "description": {
                  "type": "string",
                  "example": "this is working today"
                },
                "type": {
                  "type": "string",
                  "example": "in-progress | todo | done"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/board/task/{id}": {
      "delete": {
        "tags": [
          "Todo"
        ],
        "summary": "DELETE TODO FROM LIST",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}